some updates related to floatness of detector position

+++ czelta.pxd
@@ -13,14 +13,14 @@ cdef extern from "station.h" nogil:
         inline char* name()
         short* lastTDCCorrect()
         short* TDCCorrect(int timestamp)
-        double* detectorPosition()
+        float* detectorPosition()
         double* GPSPosition()
         
         double distanceTo(Station& st)
         
         void setName(char* name)
         void setGPSPosition(double latitude, double longitude, double height)
-        void setDetectorPosition(double x1, double y1, double x2, double y2)
+        void setDetectorPosition(float x1, float y1, float x2, float y2)
         void clearTDCCorrect(int capacity)
         void pushTDCCorrect(int fr, short tdc0, short tdc1, short tdc2)
         void pushTDCCorrect(string fr, short tdc0, short tdc1, short tdc2)
